Add files via upload
''' For this problem, use the dictionary from the beginning of this chapter whose keys are month names and whose values are the number of days in the corresponding months. (a) Ask the user to enter a month name and use the dictionary to tell them how many days are in the month. (b) Print out all of the keys in alphabetical order. (c) Print out all of the months with 31 days. (d) Print out the (key-value) pairs sorted by the number of days in each month 11.5. EXERCISES 105 (e) Modify the program from part (a) and the dictionary so that the user does not have to know how to spell the month name exactly. That is, all they have to do is spell the first three letters of the month name correctly'''